pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "workflow-monitor"
version = "0.1.0"
description = "Real-time Pegasus WMS workflow monitoring"
requires-python = ">=3.9"
dependencies = [
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
htcondor = ["htcondor"]

[project.scripts]
workflow-monitor = "workflow_monitor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/workflow_monitor"]

